move internals to parent components
