+++ src/components/CheckboxInput/CheckableInput/CheckableInput.test.tsx
